requirements.txt
fastapi==0.95.0
langchain==0.0.123
pydantic==1.10.7
uvicorn==0.21.1
pinecone-client
openai
black
isort
websockets
jinja2
bs4
unstructured
libmagic
promptlayer
